pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "pygha"
version = "0.1.0"
description = "A Python-native CI/CD framework for defining, testing, and transpiling pipelines."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "ruamel.yaml>=0.18.0",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
docs = [
    "sphinx>=8.0.2",
    "sphinx-autodoc-typehints>=2.4.4",
]

[project.scripts]
pygha = "pygha.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

# If your code lives in src/pygha/, tell hatchling to package it.
[tool.hatch.build.targets.wheel]
packages = ["src/pygha"]

# Keep tox config tiny using legacy_tox_ini so we don't duplicate sections.
[tool.tox]
requires = ["tox>=4.18"]
legacy_tox_ini = """
[tox]
envlist = py311,py312,py313,py314,lint,typecheck
skip_missing_interpreters = true

[testenv]
deps =
    pytest>=8
    pytest-cov>=5
commands = pytest --cov=pygha --cov-report=term --cov-report=xml

[testenv:lint]
deps = ruff>=0.6
commands = ruff check .

[testenv:typecheck]
deps = mypy>=1.11
commands = mypy src/pygha
"""

# MyPy config (as you added)
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
show_error_codes = true
pretty = true


[tool.coverage.run]
branch = true
source = ["pygha"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
  "if TYPE_CHECKING:",
  "pragma: no cover",
  "if __name__ == .__main__.:",
]
